feat: introduce mock module
Prior to this change, I was frustrated with exiting mocking frameworks
that have numerous and complicated options, and cumbersome APIs.  I
wanted a mock that will accept simple functions that will be used as
delegate methods.

This change introduces a set of helper functions that can be used to
satisfy an interface method by writing a single line implementation.
Each test can be given zero or more delegate functions which must
execute successfully in the context of a testing.TB value.