fix: fail the test in doCall as necessary
Prior to this change, when doCall detected a type mismatch it failed to
Fail the mock object's TB value.  Also, it didn't detect when the
delegate call returned a different number of result values than
expected.

This change performs the result value length check and moves the error
handling to outside the loop.  The error handling now calls Error on the
testing.TB value.