feat: accept optional call count for ExpectMany
Prior to this change, ExpectMany delegate would panic on repeated calls
with nil pointer dereference.  Also, it was confusing and unintuitive to
require the first argument to be an int for the call counter.

This change introduces a type for the call counter argument to
disambiguate between a regular int argument and the now optional call
count argument.