feat: introduce mockgen
Prior to this change, mocks are expected to be so simple that they may
be written by hand.  However such an exercise is tedious when faced with
many methods to be implemented.

This change introduces a mockgen command that can be used with go
generate.  The input to mockgen is a code file (or set of code files)
that have the special build tag: `mockstub`.  This approach avoids a
command with copious arguments that are hard to remember.  Instead you
simply write a struct with the interfaces that you want satisfied, and
mockgen will generate a file named mock_gen.go with the implementation
for the methods and the Expect helpers. E.g.

    //go:build mockstub
    package name

    type mockCallItWhatYouLike struct {
    	SomeInterface
    	AnotherInterface
    }