test(mockgen): introduce tests for gen subcommand
Prior to this change, the basic command line usage had been tested but
the behaviour of the gen command still had no test coverage.

This change introduces tests using the txtar format.  Each .txt file
contains a self contained module and golden files for the stdout, stderr
and the generated mock_gen.go file.  The genCmd struct is moved to an
internal package for easy testing and executed in the test process for
accurate code coverage.  Only a coulpe of basic tests are introduced.