fix(mockgen): accept flags for the fallback command
Prior to this change, the flags were never set on the genCmd when it is
invoked as the fallback command.

This change ensures that when the gen command is executed as the
fallback command, the flags are set correctly.  Also, the default logger
is decoupled from the genCmd for easier testing, the header flag is
simplified and a battery of tests have been added using the txtar
format.