refactor: use go tool friendly path for golden files
Prior to this change golden files in the txtar files had inconsistent
naming, either *.golden or simply golden.  These naming conventions
didn't clearly communicate what the file was to be compared to.

This change, renames these files so that the base name matches the
actual file that is generated and places them under testdata so that
they will be ignored by the go toolchain.