fix: remove use of os.Chdir in tests
Prior to this change, TestGenerate was using os.Chdir but this would
fail intermittently with an error that the txtar files could not be
found.  Presumably testscript relies on the working directory to find
these files.

This change introduces the GenerateOption type that can be passed to the
Execute method.  These follow the functional options pattern and allows
the test command to specify the working directory.