feat: introduce github actions and badges
Prior to this change, there was no automated checks on pull requests and
we didn't promote the quality of the project.

This change introduces github actions to test, vet and report on code
coverage.  Also, we link to the go reference page and show badges for
code quality and coverage in the README.