feat(mockgen): record code coverage for test scripts
Prior to this change, the test scripts (under ./cmd/mockgen) would run
the command using go-run, which builds the mockgen binary and then runs
it.  This meant that no code coverage was collected for these tests.

This change introduces a custom TestMain function that builds the binary
with the -cover option (new in go 1.20) which produces coverage data in
a binary format.  The go-test command can also produce data in the same
format when given the -test.gocoverdir option.  When the ./cmd/mockgen
tests find this option, coverage for the script tests is enabled.  This
allows us to then combine the coverage from the tests across all
packages.  At this time, the go-test command doesn't appear to be able
to do this automatically and the coverage reported is inaccurate.  To
generate the combined coverage use the covdata tool after running
go-test:

    mkdir "$PWD/covdata"
    go test -cover ./... -test.gocoverdir="$PWD/covdata"
    go tool covdata percent -i="$PWD/covdata"

The same tool can also convert to the old text format:

    go tool covdata textfmt -o cover.out -i="$PWD/covdata"

See https://go.dev/doc/build-cover