chore: rename package to vermock/vermockgen
Prior to this change, this package had been given the same name as an
other package which is already in widespread use within the community.
This does not cause a problem per se, but it does lead to aliasing
should a project choose use both or adopt one and migrate away from the
other.  This aliasing can lead to ambiguity when a naming convention
isn't agreed upon.  The problem can be even more difficult to diagnose
when you consider multiple installed binaries with the same name.

This change could recommend a naming convention to address such
situations but we have decided to take the more drastic action of
renaming the whole package while we are still pre-v1.  This mean
renaming the module, GitHub repository, the packages (root package,
./cmd/mockgen, ./internal/cmd/mockgen) and of course the README.