chore: improve readability of test in GH actions
Prior to this change, the test results are long and hard to see when one
test ends and another starts.

This change introduces gotestfmt tool which format go test output into
coloured, collapsible sections.  Also, the original JSON output is uploaded in
case of any issues, along with the raw coverage data.