chore: introduce fgrosse/go-coverage-report
Prior to this change, the existing code coverage report step in the
GitHub actions placed the report on the wiki, while this is good for a
historical view it doesn't provide any immediate feedback during the
pull request review process.

This change introduces fgrosse/go-coverage-report, which is capable of
participating in the conversation of the pull request.